Allow devs to skip check on specific lines
Observed problem:
In this PR: https://github.com/vikingco/mvne-platform/pull/7340/files#diff-444e80f30bd2f1e25b779966c9621b9bR285 we
wanted to bump the version of a direct dependency to a pre-build package. This PR got rejected because it did not match
our regex. This is however allowed by the PEP standard as seen here:
https://www.python.org/dev/peps/pep-0440/#direct-references

Proposed solution:
Allow devs to add a trailing ` #no-qa` to a dependency to skip the check. This should be used sparsely and thoroughly
checked during code review that there is no alternative way to add the dependency